auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Authentication service"
authors = ["Abhijeet Sasmal <abhijeetsasmal1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
